Extract GraphQL "mutation with long name" tests into its own file
It's the createOrUpdateProfileWithVerifiedPersonalInformation mutation
tests.

The one giant GraphQL test file has grown too big. It's time to split it
into smaller pieces.
